data_base/requirements.txt
# ==============================================================================
# GENERALIZED MULTI-DATABASE BACKEND SYSTEM - REQUIREMENTS
# ==============================================================================
# FastAPI SOTA Architecture with SQLite, PostgreSQL, MongoDB support
# Version: 1.0.0
# ==============================================================================

# ------------------------------------------------------------------------------
# FASTAPI CORE - High-performance async web framework
# ------------------------------------------------------------------------------
fastapi==0.109.0                    # Modern async web framework with auto-docs
uvicorn[standard]==0.27.0           # Lightning-fast ASGI server with uvloop
pydantic==2.5.3                     # Data validation using Python type hints
pydantic-settings==2.1.0            # Settings management with .env support

# ------------------------------------------------------------------------------
# DATABASE - SQL (PostgreSQL & SQLite with async support)
# ------------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.25         # ORM with native async support
asyncpg==0.29.0                     # PostgreSQL async driver (fastest)
aiosqlite==0.19.0                   # SQLite async driver for development
alembic==1.13.1                     # Database migration tool
greenlet==3.0.3                     # Required for SQLAlchemy async

# ------------------------------------------------------------------------------
# DATABASE - NOSQL (MongoDB with async Motor driver)
# ------------------------------------------------------------------------------
motor==3.3.2                        # Async MongoDB driver built on PyMongo
pymongo==4.6.1                      # MongoDB Python driver (Motor dependency)

# ------------------------------------------------------------------------------
# SECURITY - JWT Authentication & Password Hashing
# ------------------------------------------------------------------------------
python-jose[cryptography]==3.3.0    # JWT token creation and verification
bcrypt==4.0.1                       # Pin version for passlib compatibility
passlib[bcrypt]==1.7.4              # Secure password hashing with bcrypt
python-multipart==0.0.6             # Form data parsing for file uploads
cryptography==41.0.7                # Cryptographic primitives

# ------------------------------------------------------------------------------
# HTTP CLIENT - Async HTTP requests
# ------------------------------------------------------------------------------
httpx==0.26.0                       # Modern async HTTP client
aiohttp==3.9.1                      # Alternative async HTTP for webhooks

# ------------------------------------------------------------------------------
# SERIALIZATION & UTILITIES
# ------------------------------------------------------------------------------
orjson==3.9.12                      # Fast JSON serialization (10x faster)
python-dotenv==1.0.0                # Environment variable management
email-validator==2.1.0.post1        # Email validation for Pydantic
python-dateutil==2.8.2              # Date/time utilities

# ------------------------------------------------------------------------------
# CACHING & RATE LIMITING
# ------------------------------------------------------------------------------
redis==5.0.1                        # Redis client for caching
aioredis==2.0.1                     # Async Redis client

# ------------------------------------------------------------------------------
# LOGGING & MONITORING
# ------------------------------------------------------------------------------
structlog==24.1.0                   # Structured logging
loguru==0.7.2                       # Simple yet powerful logging

# ------------------------------------------------------------------------------
# TESTING - Comprehensive test suite
# ------------------------------------------------------------------------------
pytest==7.4.4                       # Testing framework
pytest-asyncio==0.23.3              # Async test support
pytest-cov==4.1.0                   # Code coverage
pytest-xdist==3.5.0                 # Parallel test execution
httpx==0.26.0                       # HTTP testing client
factory-boy==3.3.0                  # Test fixtures factory

# ------------------------------------------------------------------------------
# DEVELOPMENT TOOLS - Code quality & formatting
# ------------------------------------------------------------------------------
black==24.1.0                       # Code formatter
isort==5.13.2                       # Import sorter
mypy==1.8.0                         # Static type checker
ruff==0.1.13                        # Fast Python linter
pre-commit==3.6.0                   # Git hooks for code quality
